重构 BIT 类并新增 L218 天际线算法
重构了 L327.h 中的 BIT 类，优化了成员变量命名、方法实现及区间和统计逻辑。
新增 L218 类及其天际线计算算法，支持建筑物轮廓的动态更新。
更新了项目配置文件，添加 L218.cpp 和 L218.h 的引用。
在 main.cpp 中新增 L218 类的测试代码，调整部分测试逻辑。

+++ Lee/LCR170.h
@@ -57,7 +57,7 @@ public:
 
         BITA bit(record.size() + 1);
 
-        for (int i = record.size() - 1; i >= 0; i--) {
+        for (size_t i = record.size() - 1; i >= 0; i--) {
             ans += bit.query(record.at(i) - 1);
             bit.update(record.at(i),1);
         }
